REM-844: create testing and mock folders, move stub to it
