+++ testing/mocks/stub.py
